+++ service/savingsplans/api.go
@@ -786,7 +786,7 @@ type CreateSavingsPlanInput struct {
 	ClientToken *string `locationName:"clientToken" type:"string" idempotencyToken:"true"`
 
 	// The hourly commitment, in USD. This is a value between 0.001 and 1 million.
-	// You cannot specify more than three digits after the decimal point.
+	// You cannot specify more than five digits after the decimal point.
 	//
 	// Commitment is a required field
 	Commitment *string `locationName:"commitment" type:"string" required:"true"`